Build a tRNA abundance count matrix from charging data.

Pivot charging data into an integer matrix suitable for DESeq2, where each row is a tRNA and each column is a sample. Values are total abundance (charged + uncharged counts).

Usage

abundance_count_matrix(charging_data, min_count = 10)

Arguments

charging_data

A tibble of combined charging data with columns ref, counts_charged, counts_uncharged, and sample_id (as returned by read_charging_multi()).

min_count

Minimum total count across all samples for a tRNA to be retained. Default 10.

Value

An integer matrix with tRNA names as row names and sample IDs as column names.

Examples

results <- read_pipeline_results(
  clover_example("ecoli/config.yaml"),
  types = "charging"
)
abundance_count_matrix(results$charging)
#>                                  wt-15-ctl-01 wt-15-ctl-02 wt-15-ctl-03
#> host-tRNA-Ala-GGC-1-1                     298          397          113
#> host-tRNA-Ala-GGC-1-1-uncharged          4393         5014          883
#> host-tRNA-Ala-GGC-1-2                    1625         1506          265
#> host-tRNA-Ala-GGC-1-2-uncharged          4391         5130          862
#> host-tRNA-Ala-TGC-1-1                     518          610          160
#> host-tRNA-Ala-TGC-1-1-uncharged          3785         5079         1051
#> host-tRNA-Ala-TGC-1-2                     829          808          189
#> host-tRNA-Ala-TGC-1-2-uncharged         17942        17266         2182
#> host-tRNA-Ala-TGC-1-3                    2329         2191          355
#> host-tRNA-Ala-TGC-1-3-uncharged          3651         5111          937
#> host-tRNA-Arg-ACG-1-1                    1414         1211          308
#> host-tRNA-Arg-ACG-1-1-uncharged         10398        10664         2213
#> host-tRNA-Arg-ACG-1-2                    2365         1595          460
#> host-tRNA-Arg-ACG-1-2-uncharged         10285        10641         2236
#> host-tRNA-Arg-ACG-1-3                    1462         1160          343
#> host-tRNA-Arg-ACG-1-3-uncharged         12792        13839         2607
#> host-tRNA-Arg-ACG-1-4                    1718         1401          379
#> host-tRNA-Arg-ACG-1-4-uncharged         10108        10459         2188
#> host-tRNA-Arg-CCG-1-1                     265          395           85
#> host-tRNA-Arg-CCG-1-1-uncharged          1483         1798          339
#> host-tRNA-Arg-CCT-1-1                     183          248           39
#> host-tRNA-Arg-CCT-1-1-uncharged           601         1474          234
#> host-tRNA-Arg-TCT-1-1                     397          284          115
#> host-tRNA-Arg-TCT-1-1-uncharged          2355         1995          473
#> host-tRNA-Asn-GTT-1-1                     404          408          149
#> host-tRNA-Asn-GTT-1-1-uncharged          1948         2661          786
#> host-tRNA-Asn-GTT-1-2                     284          281          126
#> host-tRNA-Asn-GTT-1-2-uncharged          2135         2721          771
#> host-tRNA-Asn-GTT-1-3                     412          378          130
#> host-tRNA-Asn-GTT-1-3-uncharged          2043         2744          830
#> host-tRNA-Asn-GTT-1-4                     388          384          152
#> host-tRNA-Asn-GTT-1-4-uncharged         17489        19662         3161
#> host-tRNA-Asp-GTC-1-1                    1446         1790         1196
#> host-tRNA-Asp-GTC-1-1-uncharged          6153         7521         2884
#> host-tRNA-Asp-GTC-1-2                    3312         4013         1715
#> host-tRNA-Asp-GTC-1-2-uncharged          6002         7483         2837
#> host-tRNA-Asp-GTC-1-3                    3324         4068         1737
#> host-tRNA-Asp-GTC-1-3-uncharged          6183         7619         2813
#> host-tRNA-Cys-GCA-1-1                    1144         1252          577
#> host-tRNA-Cys-GCA-1-1-uncharged          9652        22378         3782
#> host-tRNA-Gln-CTG-1-1                    1093          809          395
#> host-tRNA-Gln-CTG-1-1-uncharged          5009         5751         1385
#> host-tRNA-Gln-CTG-1-2                    2786         2736          569
#> host-tRNA-Gln-CTG-1-2-uncharged          4974         5610         1346
#> host-tRNA-Gln-TTG-1-1                     930          755          390
#> host-tRNA-Gln-TTG-1-1-uncharged          8930         9915         2330
#> host-tRNA-Gln-TTG-1-2                    1296         1014          493
#> host-tRNA-Gln-TTG-1-2-uncharged          3273         3511         1258
#> host-tRNA-Glu-TTC-1-1                    4354         5104         1136
#> host-tRNA-Glu-TTC-1-1-uncharged          2254         2857         1106
#> host-tRNA-Glu-TTC-1-2                    4478         4920         1166
#> host-tRNA-Glu-TTC-1-2-uncharged          2272         2896         1116
#> host-tRNA-Glu-TTC-1-3                    1167         1229          595
#> host-tRNA-Glu-TTC-1-3-uncharged          6620         7258         1719
#> host-tRNA-Glu-TTC-1-4                    1216         1244          606
#> host-tRNA-Glu-TTC-1-4-uncharged         12217        14467         2315
#> host-tRNA-Gly-CCC-1-1                     315          390          193
#> host-tRNA-Gly-CCC-1-1-uncharged          2956         4926          959
#> host-tRNA-Gly-GCC-1-1                     847          853          440
#> host-tRNA-Gly-GCC-1-1-uncharged         10699        15983         3373
#> host-tRNA-Gly-GCC-1-2                     866          871          434
#> host-tRNA-Gly-GCC-1-2-uncharged         10778        15520         3250
#> host-tRNA-Gly-GCC-1-3                     663          832          440
#> host-tRNA-Gly-GCC-1-3-uncharged         10893        16073         3368
#> host-tRNA-Gly-GCC-1-4                    2149         2676          807
#> host-tRNA-Gly-GCC-1-4-uncharged          4350         7284         2406
#> host-tRNA-Gly-TCC-1-1                     960         1245          404
#> host-tRNA-Gly-TCC-1-1-uncharged          6275         9090         1781
#> host-tRNA-His-GTG-1-1                    1182         1154          476
#> host-tRNA-His-GTG-1-1-uncharged          9593        11585         2434
#> host-tRNA-Ile-GAT-1-1                    2587         3353         8446
#> host-tRNA-Ile-GAT-1-1-uncharged         11402        15346        22236
#> host-tRNA-Ile-GAT-1-2                    2132         2752         7240
#> host-tRNA-Ile-GAT-1-2-uncharged         11087        14972        21708
#> host-tRNA-Ile-GAT-1-3                    2575         3397         8516
#> host-tRNA-Ile-GAT-1-3-uncharged         11380        15087        22204
#> host-tRNA-Ile2-CAT-1-1                    561          736         1213
#> host-tRNA-Ile2-CAT-1-1-uncharged         1177         1477         1681
#> host-tRNA-Ile2-CAT-1-2                     68           70          131
#> host-tRNA-Ile2-CAT-1-2-uncharged          160          344          257
#> host-tRNA-Leu-CAA-1-1                    1375         1792          352
#> host-tRNA-Leu-CAA-1-1-uncharged          3416         5383          846
#> host-tRNA-Leu-CAG-1-1                    1331         1462          362
#> host-tRNA-Leu-CAG-1-1-uncharged          6177         8020         1741
#> host-tRNA-Leu-CAG-1-2                    1470         1323          382
#> host-tRNA-Leu-CAG-1-2-uncharged         14096        14284         2040
#> host-tRNA-Leu-CAG-1-3                     808          978          330
#> host-tRNA-Leu-CAG-1-3-uncharged          4909         6241         1622
#> host-tRNA-Leu-CAG-1-4                    4136         3842          496
#> host-tRNA-Leu-CAG-1-4-uncharged         14299        14516         2088
#> host-tRNA-Leu-GAG-1-1                     942         1109          265
#> host-tRNA-Leu-GAG-1-1-uncharged          8118        10718         1732
#> host-tRNA-Leu-TAA-1-1                    1719         2062          611
#> host-tRNA-Leu-TAA-1-1-uncharged          7825        10883         2220
#> host-tRNA-Leu-TAG-1-1                    1995         1918          437
#> host-tRNA-Leu-TAG-1-1-uncharged          5033         5245         1259
#> host-tRNA-Lys-TTT-1-1                     831          799          538
#> host-tRNA-Lys-TTT-1-1-uncharged          6171         5512         1786
#> host-tRNA-Lys-TTT-1-2                    1952         1297          835
#> host-tRNA-Lys-TTT-1-2-uncharged          2121         2294          910
#> host-tRNA-Lys-TTT-1-3                     890          819          516
#> host-tRNA-Lys-TTT-1-3-uncharged          6045         5443         1897
#> host-tRNA-Lys-TTT-1-4                     683          607          380
#> host-tRNA-Lys-TTT-1-4-uncharged          6113         5567         1760
#> host-tRNA-Lys-TTT-1-5                     645          627          375
#> host-tRNA-Lys-TTT-1-5-uncharged          2181         2313          910
#> host-tRNA-Lys-TTT-1-6                     676          599          324
#> host-tRNA-Lys-TTT-1-6-uncharged          3569         3536         1599
#> host-tRNA-Met-CAT-1-1                    1397         1354          655
#> host-tRNA-Met-CAT-1-1-uncharged          7632         8773         2102
#> host-tRNA-Met-CAT-1-2                    1414         1322          598
#> host-tRNA-Met-CAT-1-2-uncharged          3124         3830         1453
#> host-tRNA-Phe-GAA-1-1                    2351         2876          863
#> host-tRNA-Phe-GAA-1-1-uncharged          2001         2985          846
#> host-tRNA-Phe-GAA-1-2                     503          778          416
#> host-tRNA-Phe-GAA-1-2-uncharged          2068         2978         1057
#> host-tRNA-Pro-CGG-1-1                    1528         1459          353
#> host-tRNA-Pro-CGG-1-1-uncharged         10789        15539         2880
#> host-tRNA-Pro-GGG-1-1                     302          319           58
#> host-tRNA-Pro-GGG-1-1-uncharged          1867         2582          396
#> host-tRNA-Pro-TGG-1-1                    1212         1166          313
#> host-tRNA-Pro-TGG-1-1-uncharged         12187        13737         2015
#> host-tRNA-SeC-TCA-1-1                      44           52           31
#> host-tRNA-SeC-TCA-1-1-uncharged          1203         5655          821
#> host-tRNA-Ser-CGA-1-1                     109          135          301
#> host-tRNA-Ser-CGA-1-1-uncharged          1495         4108         1917
#> host-tRNA-Ser-GCT-1-1                    1844         2429         1377
#> host-tRNA-Ser-GCT-1-1-uncharged          9363        11587         3636
#> host-tRNA-Ser-GGA-1-1                    1294         1064          586
#> host-tRNA-Ser-GGA-1-1-uncharged          5680         5760         1672
#> host-tRNA-Ser-GGA-1-2                     283          560          421
#> host-tRNA-Ser-GGA-1-2-uncharged          3082         4711         1608
#> host-tRNA-Ser-TGA-1-1                     972         1275          855
#> host-tRNA-Ser-TGA-1-1-uncharged          6956         8504         2874
#> host-tRNA-Thr-CGT-1-1                     526          623          591
#> host-tRNA-Thr-CGT-1-1-uncharged          2196         4671         1568
#> host-tRNA-Thr-CGT-1-2                       6           11          227
#> host-tRNA-Thr-CGT-1-2-uncharged           340         1931          756
#> host-tRNA-Thr-GGT-1-1                    1320         1155         2287
#> host-tRNA-Thr-GGT-1-1-uncharged          5373         8810         4505
#> host-tRNA-Thr-GGT-1-2                    1520         1226         2481
#> host-tRNA-Thr-GGT-1-2-uncharged          8425        10975         5763
#> host-tRNA-Thr-TGT-1-1                    2829         2962         3106
#> host-tRNA-Thr-TGT-1-1-uncharged          5893         8405         4938
#> host-tRNA-Trp-CCA-1-1                    1475         2254         1518
#> host-tRNA-Trp-CCA-1-1-uncharged          9567        30025         7063
#> host-tRNA-Tyr-GTA-1-1                    1896         1942          608
#> host-tRNA-Tyr-GTA-1-1-uncharged          8859        10835         2297
#> host-tRNA-Tyr-GTA-1-2                     646          698          528
#> host-tRNA-Tyr-GTA-1-2-uncharged          3179         4374         1771
#> host-tRNA-Tyr-GTA-1-3                     748          749          567
#> host-tRNA-Tyr-GTA-1-3-uncharged          3263         4373         1900
#> host-tRNA-Val-GAC-1-1                     494          727         1256
#> host-tRNA-Val-GAC-1-1-uncharged          3122         4556         3419
#> host-tRNA-Val-GAC-1-2                    1006         1287         2426
#> host-tRNA-Val-GAC-1-2-uncharged          3732         4447         5318
#> host-tRNA-Val-TAC-1-1                    1390         1290         1110
#> host-tRNA-Val-TAC-1-1-uncharged          1693         2929         2217
#> host-tRNA-Val-TAC-1-2                     574          546          969
#> host-tRNA-Val-TAC-1-2-uncharged          3274         4271         2426
#> host-tRNA-Val-TAC-1-3                     543          537          960
#> host-tRNA-Val-TAC-1-3-uncharged          3245         4307         2475
#> host-tRNA-Val-TAC-1-4                     451          414          856
#> host-tRNA-Val-TAC-1-4-uncharged          1450         2617         2119
#> host-tRNA-Val-TAC-1-5                     523          475          948
#> host-tRNA-Val-TAC-1-5-uncharged          1507         2511         2202
#> host-tRNA-fMet-CAT-1-1                    670          503          198
#> host-tRNA-fMet-CAT-1-1-uncharged         9125         9696         1798
#> host-tRNA-fMet-CAT-1-2                    618          417          152
#> host-tRNA-fMet-CAT-1-2-uncharged         6540         7277         1456
#> host-tRNA-fMet-CAT-1-3                    695          589          206
#> host-tRNA-fMet-CAT-1-3-uncharged        13176        13625         2355
#> host-tRNA-fMet-CAT-1-4                    536          439          199
#> host-tRNA-fMet-CAT-1-4-uncharged         7013         8754         1887
#> phage-tRNA-Arg-TCT                          8            3            2
#> phage-tRNA-Arg-TCT-uncharged              472         1646          230
#> phage-tRNA-Gln-TTG                         22           39           18
#> phage-tRNA-Gln-TTG-uncharged             2326        15249         2052
#> phage-tRNA-Gly-TCC                         27           33          249
#> phage-tRNA-Gly-TCC-uncharged             1966        12237         1562
#> phage-tRNA-Ile2-CAT                         9            8           11
#> phage-tRNA-Ile2-CAT-uncharged             502         1989          316
#> phage-tRNA-Leu-TAA                         10            7           22
#> phage-tRNA-Leu-TAA-uncharged              307         1431          394
#> phage-tRNA-Pro-TGG                         11            9            7
#> phage-tRNA-Pro-TGG-uncharged              657         1415          335
#> phage-tRNA-Ser-TGA                          2            5           13
#> phage-tRNA-Ser-TGA-uncharged              461         2112          357
#> phage-tRNA-Thr-TGT                         12            4           28
#> phage-tRNA-Thr-TGT-uncharged              370         1642          354
#>                                  wt-15-inf-01 wt-15-inf-02 wt-15-inf-03
#> host-tRNA-Ala-GGC-1-1                     189          347         1253
#> host-tRNA-Ala-GGC-1-1-uncharged          3640         7133        12837
#> host-tRNA-Ala-GGC-1-2                    1299         1857         4409
#> host-tRNA-Ala-GGC-1-2-uncharged          3637         7411        12592
#> host-tRNA-Ala-TGC-1-1                     233          553         1706
#> host-tRNA-Ala-TGC-1-1-uncharged          3001         7304        12955
#> host-tRNA-Ala-TGC-1-2                     377          728         2580
#> host-tRNA-Ala-TGC-1-2-uncharged         14696        26056        43550
#> host-tRNA-Ala-TGC-1-3                    1689         2723         6043
#> host-tRNA-Ala-TGC-1-3-uncharged          2794         7352        13091
#> host-tRNA-Arg-ACG-1-1                     819         1687         4539
#> host-tRNA-Arg-ACG-1-1-uncharged          7766        18225        31587
#> host-tRNA-Arg-ACG-1-2                    1168         2306         6593
#> host-tRNA-Arg-ACG-1-2-uncharged          7808        18267        31848
#> host-tRNA-Arg-ACG-1-3                     859         1684         4582
#> host-tRNA-Arg-ACG-1-3-uncharged         10629        27441        44117
#> host-tRNA-Arg-ACG-1-4                    1065         2064         5553
#> host-tRNA-Arg-ACG-1-4-uncharged          7756        18207        31506
#> host-tRNA-Arg-CCG-1-1                     216          450         1072
#> host-tRNA-Arg-CCG-1-1-uncharged          1145         2907         5217
#> host-tRNA-Arg-CCT-1-1                     143          409          697
#> host-tRNA-Arg-CCT-1-1-uncharged           492         2396         2847
#> host-tRNA-Arg-TCT-1-1                     188          341          894
#> host-tRNA-Arg-TCT-1-1-uncharged          1766         3431         5680
#> host-tRNA-Asn-GTT-1-1                     322          832         1332
#> host-tRNA-Asn-GTT-1-1-uncharged          1796         4449         8021
#> host-tRNA-Asn-GTT-1-2                     265          633         1136
#> host-tRNA-Asn-GTT-1-2-uncharged          1802         4673         8112
#> host-tRNA-Asn-GTT-1-3                     330          787         1386
#> host-tRNA-Asn-GTT-1-3-uncharged          1884         4619         8147
#> host-tRNA-Asn-GTT-1-4                     311          748         1350
#> host-tRNA-Asn-GTT-1-4-uncharged         13500        36819        52433
#> host-tRNA-Asp-GTC-1-1                    1002         2253         6231
#> host-tRNA-Asp-GTC-1-1-uncharged          4831        13889        22316
#> host-tRNA-Asp-GTC-1-2                    2492         6817        13128
#> host-tRNA-Asp-GTC-1-2-uncharged          4590        13371        21800
#> host-tRNA-Asp-GTC-1-3                    2475         6772        13054
#> host-tRNA-Asp-GTC-1-3-uncharged          4848        13679        21965
#> host-tRNA-Cys-GCA-1-1                     771         1753         3211
#> host-tRNA-Cys-GCA-1-1-uncharged          7820        35533        38003
#> host-tRNA-Gln-CTG-1-1                     588          900         2923
#> host-tRNA-Gln-CTG-1-1-uncharged          3656         9860        15997
#> host-tRNA-Gln-CTG-1-2                    1986         4750         8603
#> host-tRNA-Gln-CTG-1-2-uncharged          3746         9893        15781
#> host-tRNA-Gln-TTG-1-1                     468          860         2629
#> host-tRNA-Gln-TTG-1-1-uncharged          6998        18640        28429
#> host-tRNA-Gln-TTG-1-2                     715         1255         3555
#> host-tRNA-Gln-TTG-1-2-uncharged          2267         5463        11335
#> host-tRNA-Glu-TTC-1-1                    3289         4582        14583
#> host-tRNA-Glu-TTC-1-1-uncharged          1470         2719         8837
#> host-tRNA-Glu-TTC-1-2                    3322         4612        14438
#> host-tRNA-Glu-TTC-1-2-uncharged          1486         2625         8931
#> host-tRNA-Glu-TTC-1-3                     674         1064         4757
#> host-tRNA-Glu-TTC-1-3-uncharged          4605         7333        22484
#> host-tRNA-Glu-TTC-1-4                     660         1024         4757
#> host-tRNA-Glu-TTC-1-4-uncharged          9092        12699        33682
#> host-tRNA-Gly-CCC-1-1                     186          396         1191
#> host-tRNA-Gly-CCC-1-1-uncharged          2454         8408        11163
#> host-tRNA-Gly-GCC-1-1                     343          747         2027
#> host-tRNA-Gly-GCC-1-1-uncharged          8925        30304        38373
#> host-tRNA-Gly-GCC-1-2                     337          770         2041
#> host-tRNA-Gly-GCC-1-2-uncharged          8970        29743        37845
#> host-tRNA-Gly-GCC-1-3                     303          662         1977
#> host-tRNA-Gly-GCC-1-3-uncharged          9020        29785        38164
#> host-tRNA-Gly-GCC-1-4                    1455         4431         7532
#> host-tRNA-Gly-GCC-1-4-uncharged          3483        10526        15504
#> host-tRNA-Gly-TCC-1-1                     539         1267         3753
#> host-tRNA-Gly-TCC-1-1-uncharged          4718        11827        22581
#> host-tRNA-His-GTG-1-1                     896         2210         4190
#> host-tRNA-His-GTG-1-1-uncharged          7567        23432        32560
#> host-tRNA-Ile-GAT-1-1                    2064         6088        15069
#> host-tRNA-Ile-GAT-1-1-uncharged          9234        29026        48675
#> host-tRNA-Ile-GAT-1-2                    1819         4867        12539
#> host-tRNA-Ile-GAT-1-2-uncharged          9382        28581        47964
#> host-tRNA-Ile-GAT-1-3                    1949         5983        15104
#> host-tRNA-Ile-GAT-1-3-uncharged          9334        28943        48729
#> host-tRNA-Ile2-CAT-1-1                    426         1549         2825
#> host-tRNA-Ile2-CAT-1-1-uncharged          888         3029         4832
#> host-tRNA-Ile2-CAT-1-2                     62          173          355
#> host-tRNA-Ile2-CAT-1-2-uncharged          283          706         1352
#> host-tRNA-Leu-CAA-1-1                     897         3981         4685
#> host-tRNA-Leu-CAA-1-1-uncharged          2505        14084        12745
#> host-tRNA-Leu-CAG-1-1                     276         1876         4895
#> host-tRNA-Leu-CAG-1-1-uncharged          1531        11372        18263
#> host-tRNA-Leu-CAG-1-2                     289         1499         4440
#> host-tRNA-Leu-CAG-1-2-uncharged          4179        22799        27579
#> host-tRNA-Leu-CAG-1-3                     297         1194         2865
#> host-tRNA-Leu-CAG-1-3-uncharged          1942        11180        14692
#> host-tRNA-Leu-CAG-1-4                    1159         5381         9246
#> host-tRNA-Leu-CAG-1-4-uncharged          4269        22451        27698
#> host-tRNA-Leu-GAG-1-1                     515         1309         4374
#> host-tRNA-Leu-GAG-1-1-uncharged          5879        19481        28787
#> host-tRNA-Leu-TAA-1-1                     903         3083         7178
#> host-tRNA-Leu-TAA-1-1-uncharged          5915        24317        29362
#> host-tRNA-Leu-TAG-1-1                    1358         2885         6371
#> host-tRNA-Leu-TAG-1-1-uncharged          3656        10423        16914
#> host-tRNA-Lys-TTT-1-1                     589         1192         3062
#> host-tRNA-Lys-TTT-1-1-uncharged          5795        12870        19187
#> host-tRNA-Lys-TTT-1-2                    1084         2248         5817
#> host-tRNA-Lys-TTT-1-2-uncharged          1813         5834         9445
#> host-tRNA-Lys-TTT-1-3                     559         1260         2974
#> host-tRNA-Lys-TTT-1-3-uncharged          5554        12513        18721
#> host-tRNA-Lys-TTT-1-4                     463         1087         2326
#> host-tRNA-Lys-TTT-1-4-uncharged          5777        12906        18868
#> host-tRNA-Lys-TTT-1-5                     517         1068         2293
#> host-tRNA-Lys-TTT-1-5-uncharged          1842         5730         9451
#> host-tRNA-Lys-TTT-1-6                     484         1066         2293
#> host-tRNA-Lys-TTT-1-6-uncharged          2522         7095        13552
#> host-tRNA-Met-CAT-1-1                    1047         2252         4352
#> host-tRNA-Met-CAT-1-1-uncharged          6652        18622        24814
#> host-tRNA-Met-CAT-1-2                    1059         2360         4344
#> host-tRNA-Met-CAT-1-2-uncharged          2556         7860        12609
#> host-tRNA-Phe-GAA-1-1                    1829         4494         8020
#> host-tRNA-Phe-GAA-1-1-uncharged          1663         5081         8438
#> host-tRNA-Phe-GAA-1-2                     400         1024         2122
#> host-tRNA-Phe-GAA-1-2-uncharged          1562         4806         8724
#> host-tRNA-Pro-CGG-1-1                     906         1853         3585
#> host-tRNA-Pro-CGG-1-1-uncharged          9594        24794        33025
#> host-tRNA-Pro-GGG-1-1                     158          113          676
#> host-tRNA-Pro-GGG-1-1-uncharged          1622         1622         5500
#> host-tRNA-Pro-TGG-1-1                     567         1067         3469
#> host-tRNA-Pro-TGG-1-1-uncharged         10964        22136        37706
#> host-tRNA-SeC-TCA-1-1                      10           39          120
#> host-tRNA-SeC-TCA-1-1-uncharged           954         8450         7521
#> host-tRNA-Ser-CGA-1-1                      76          175          455
#> host-tRNA-Ser-CGA-1-1-uncharged          1184         6419         6981
#> host-tRNA-Ser-GCT-1-1                    1498         3503         6094
#> host-tRNA-Ser-GCT-1-1-uncharged          6656        20859        25566
#> host-tRNA-Ser-GGA-1-1                     819         2029         3823
#> host-tRNA-Ser-GGA-1-1-uncharged          3313        11433        15376
#> host-tRNA-Ser-GGA-1-2                     298          621         1733
#> host-tRNA-Ser-GGA-1-2-uncharged          2297         8491        11976
#> host-tRNA-Ser-TGA-1-1                     746         2051         3632
#> host-tRNA-Ser-TGA-1-1-uncharged          4578        15567        21062
#> host-tRNA-Thr-CGT-1-1                     421         1250         2613
#> host-tRNA-Thr-CGT-1-1-uncharged          1751         7762         9919
#> host-tRNA-Thr-CGT-1-2                       5           26           86
#> host-tRNA-Thr-CGT-1-2-uncharged           308         2739         2713
#> host-tRNA-Thr-GGT-1-1                    1190         2680         6667
#> host-tRNA-Thr-GGT-1-1-uncharged          5023        15182        20943
#> host-tRNA-Thr-GGT-1-2                    1391         2838         7123
#> host-tRNA-Thr-GGT-1-2-uncharged          6717        19168        27673
#> host-tRNA-Thr-TGT-1-1                    2249         5596        11482
#> host-tRNA-Thr-TGT-1-1-uncharged          4798        14464        22086
#> host-tRNA-Trp-CCA-1-1                     933         3141         6702
#> host-tRNA-Trp-CCA-1-1-uncharged          8130        48380        51116
#> host-tRNA-Tyr-GTA-1-1                    1172         2695         6276
#> host-tRNA-Tyr-GTA-1-1-uncharged          6089        19576        28677
#> host-tRNA-Tyr-GTA-1-2                     407          970         2132
#> host-tRNA-Tyr-GTA-1-2-uncharged          2268         7964        11177
#> host-tRNA-Tyr-GTA-1-3                     453         1050         2527
#> host-tRNA-Tyr-GTA-1-3-uncharged          2307         7945        11159
#> host-tRNA-Val-GAC-1-1                     441         1317         3036
#> host-tRNA-Val-GAC-1-1-uncharged          2613         7984        12967
#> host-tRNA-Val-GAC-1-2                     847         1947         5453
#> host-tRNA-Val-GAC-1-2-uncharged          3135         6396        15666
#> host-tRNA-Val-TAC-1-1                    1326         2142         4900
#> host-tRNA-Val-TAC-1-1-uncharged          1676         5205         7548
#> host-tRNA-Val-TAC-1-2                     532         1080         2489
#> host-tRNA-Val-TAC-1-2-uncharged          3092         6907        10744
#> host-tRNA-Val-TAC-1-3                     490          969         2523
#> host-tRNA-Val-TAC-1-3-uncharged          3099         6999        11073
#> host-tRNA-Val-TAC-1-4                     396          758         1949
#> host-tRNA-Val-TAC-1-4-uncharged          1536         4461         6812
#> host-tRNA-Val-TAC-1-5                     537          863         2183
#> host-tRNA-Val-TAC-1-5-uncharged          1547         4583         6390
#> host-tRNA-fMet-CAT-1-1                    491         1050         1753
#> host-tRNA-fMet-CAT-1-1-uncharged         6901        21186        27608
#> host-tRNA-fMet-CAT-1-2                    426          946         1497
#> host-tRNA-fMet-CAT-1-2-uncharged         5077        17004        22517
#> host-tRNA-fMet-CAT-1-3                    505         1311         2095
#> host-tRNA-fMet-CAT-1-3-uncharged         9475        27497        36920
#> host-tRNA-fMet-CAT-1-4                    357          898         1333
#> host-tRNA-fMet-CAT-1-4-uncharged         5430        18050        21916
#> phage-tRNA-Arg-TCT                        180          129          718
#> phage-tRNA-Arg-TCT-uncharged             2998         4370         8971
#> phage-tRNA-Gln-TTG                        441          593         1307
#> phage-tRNA-Gln-TTG-uncharged             5355        27963        26165
#> phage-tRNA-Gly-TCC                        341          634         1391
#> phage-tRNA-Gly-TCC-uncharged             5564        26127        23572
#> phage-tRNA-Ile2-CAT                       501          720         1974
#> phage-tRNA-Ile2-CAT-uncharged            2599         6242         9170
#> phage-tRNA-Leu-TAA                        241          452         1339
#> phage-tRNA-Leu-TAA-uncharged             2924         7488        10550
#> phage-tRNA-Pro-TGG                        109          137          406
#> phage-tRNA-Pro-TGG-uncharged             2863         5846         9340
#> phage-tRNA-Ser-TGA                        153          283          643
#> phage-tRNA-Ser-TGA-uncharged             1674         5941         7552
#> phage-tRNA-Thr-TGT                        550          796         1799
#> phage-tRNA-Thr-TGT-uncharged             2736         6123         7923